requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Async HTTP
httpx>=0.27.0
aiohttp>=3.9.0

# Auth / security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
asyncpg>=0.29.0

# Cache
redis>=5.0.0

# ── Google AI ──────────────────────────────────────────────────────────────
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0
google-cloud-vision>=3.7.0
google-cloud-storage>=2.17.0

# ── Other AI / multimodal ──────────────────────────────────────────────────
openai>=1.30.0
anthropologic>=0.28.0

# ── Image processing ───────────────────────────────────────────────────────
pillow>=10.3.0
opencv-python-headless>=4.10.0
imagehash>=4.3.1
rembg>=2.0.57
pytesseract>=0.3.10
scikit-image>=0.23.0

# ── eBay SDK ───────────────────────────────────────────────────────────────
ebaysdk>=2.2.0

# ── Telegram ───────────────────────────────────────────────────────────────
aiogram>=3.7.0
python-telegram-bot>=21.0

# ── Data / utilities ───────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2

# ── Dev / test ─────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.4.0
